+++ src/containers/PetsList.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import PetCard from './PetCard';
+import PetCard from '../components/PetCard';
 import Pagination from './Pagination';
 import FilterParams from './FilterParams';
 import { getPetsList } from '../actions/PetActions';
