+++ src/test/components/Bottom.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { render } from '@testing-library/react';
-import Bottom from '../../containers/Bottom';
+import Bottom from '../../components/Bottom';
 
 it('Renders without crashing', () => {
   const div = document.createElement('div');
