+++ src/test/components/RouteDisplayer.test.js
@@ -4,13 +4,17 @@ import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { render, cleanup } from '@testing-library/react';
-import FilterParams from '../../components/FilterParams';
+import RouteDisplayer from '../../components/RouteDisplayer';
 
 afterEach(cleanup);
 
+const initialState = {
+  pets: {},
+};
+
 const composeEnhancers = compose;
 
-const reducer = (state, action) => {
+const reducer = (state = initialState, action) => {
   switch (action.type) {
     case 'CHANGE_PET_FILTER':
       return {
@@ -21,7 +25,7 @@ const reducer = (state, action) => {
   }
 };
 
-const defaultStore = createStore(reducer, null, composeEnhancers(applyMiddleware(thunk)));
+const defaultStore = createStore(reducer, initialState, composeEnhancers(applyMiddleware(thunk)));
 
 const renderWithRedux = (component, { store = defaultStore } = {}) => ({
   ...render(
@@ -32,6 +36,6 @@ const renderWithRedux = (component, { store = defaultStore } = {}) => ({
 });
 
 it('Renders using Redux without crashing', () => {
-  const filterParams = renderWithRedux(<FilterParams />);
-  expect(filterParams).toBeTruthy();
+  const routeDisplayer = renderWithRedux(<RouteDisplayer />);
+  expect(routeDisplayer).toBeTruthy();
 });
