+++ src/test/containers/FilterParams.test.js
@@ -4,42 +4,16 @@ import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { render, cleanup } from '@testing-library/react';
-import PetsList from '../../components/PetsList';
+import FilterParams from '../../containers/FilterParams';
 
 afterEach(cleanup);
-
-const initialState = {
-  pets: {
-    animals: [
-      {
-        id: 147852369,
-        photos: [],
-        name: 'Chewbaca',
-        published_at: '2020-10-25',
-        breeds: {
-          mixed: true,
-        },
-        contact: {
-          email: '',
-          phone: '',
-          address: {
-            city: '',
-            state: '',
-          },
-        },
-      },
-    ],
-    pagination: {},
-  },
-};
 
 const composeEnhancers = compose;
 
 const reducer = (state, action) => {
   switch (action.type) {
-    case 'GET_PETS_LIST':
+    case 'CHANGE_PET_FILTER':
       return {
-        ...state,
         ...action.payload,
       };
     default:
@@ -47,11 +21,7 @@ const reducer = (state, action) => {
   }
 };
 
-const defaultStore = createStore(reducer,
-  initialState,
-  composeEnhancers(
-    applyMiddleware(thunk),
-  ));
+const defaultStore = createStore(reducer, null, composeEnhancers(applyMiddleware(thunk)));
 
 const renderWithRedux = (component, { store = defaultStore } = {}) => ({
   ...render(
@@ -62,6 +32,6 @@ const renderWithRedux = (component, { store = defaultStore } = {}) => ({
 });
 
 it('Renders using Redux without crashing', () => {
-  const petsList = renderWithRedux(<PetsList />);
-  expect(petsList).toBeTruthy();
+  const filterParams = renderWithRedux(<FilterParams />);
+  expect(filterParams).toBeTruthy();
 });
