+++ src/test/containers/PetDetails.test.js
@@ -4,20 +4,39 @@ import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { render, cleanup } from '@testing-library/react';
-import RouteDisplayer from '../../containers/RouteDisplayer';
+import PetDetails from '../../containers/PetDetails';
 
 afterEach(cleanup);
 
 const initialState = {
-  pets: {},
+  pets: {
+    animal: {
+      id: 147852369,
+      photos: [],
+      name: 'Chewbaca',
+      published_at: '2020-10-25',
+      breeds: {
+        mixed: true,
+      },
+      contact: {
+        email: '',
+        phone: '',
+        address: {
+          city: '',
+          state: '',
+        },
+      },
+    },
+  },
 };
 
 const composeEnhancers = compose;
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
-    case 'CHANGE_PET_FILTER':
+    case 'GET_PET_INFO':
       return {
+        ...state,
         ...action.payload,
       };
     default:
@@ -25,7 +44,11 @@ const reducer = (state = initialState, action) => {
   }
 };
 
-const defaultStore = createStore(reducer, initialState, composeEnhancers(applyMiddleware(thunk)));
+const defaultStore = createStore(reducer,
+  initialState,
+  composeEnhancers(
+    applyMiddleware(thunk),
+  ));
 
 const renderWithRedux = (component, { store = defaultStore } = {}) => ({
   ...render(
@@ -36,6 +59,11 @@ const renderWithRedux = (component, { store = defaultStore } = {}) => ({
 });
 
 it('Renders using Redux without crashing', () => {
-  const routeDisplayer = renderWithRedux(<RouteDisplayer />);
-  expect(routeDisplayer).toBeTruthy();
+  const match = {
+    params: {
+      id: '147852369',
+    },
+  };
+  const petDetails = renderWithRedux(<PetDetails match={match} />);
+  expect(petDetails).toBeTruthy();
 });
